update ed ativ2

+++ SEMESTRES/3/ED/Ativ2/include/operation.hpp
@@ -9,16 +9,16 @@ typedef long long ll;
 
 struct OperationReturn {
   ll value;
-  float user_time;
-  float sys_time;
-  float clock_time;
+  double user_time;
+  double sys_time;
+  double clock_time;
 };
 
 class Operation {
   public:
-    float get_sys_time_exec(rusage end, rusage start);
-    float get_user_time_exec(rusage end, rusage start);
-    float get_clock_time_exec(timespec end, timespec start);
+    double get_sys_time_exec(rusage end, rusage start);
+    double get_user_time_exec(rusage end, rusage start);
+    double get_clock_time_exec(timespec end, timespec start);
 
     Operation();
     OperationReturn recursion(ll n);
